CoachEd_flashhired/requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
Babel==2.9.1
blis==0.7.4
buildozer==1.2.0
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cymem==2.0.5
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
django-allauth==0.44.0
django-bootstrap-form==3.4
django-extensions==3.1.3
django-filter==2.4.0
django-phonenumber-field==5.2.0
django-registration==3.1.1
django-restframework==0.0.1
django-survey-and-report==1.3.31
django-tastypie==0.14.3
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djangotoolbox==1.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.1.0
docutils==0.17.1
filelock==3.0.12
Flask==2.0.1
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
isort==5.6.4
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.0
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.2
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
KivyAuth==2.3.2
kivymd==0.104.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
mathparse==0.1.2
mccabe==0.6.1
mongoengine==0.23.1
monotonic==1.6
murmurhash==1.0.5
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.5.2.52
ordereddict==1.1
packaging==20.8
pefile==2021.5.13
pexpect==4.8.0
phonenumbers==8.12.31
Pillow==8.3.2
Pint==0.16.1
plac==1.1.3
plyer==2.0.0
preshed==3.0.5
ptyprocess==0.7.0
pycparser==2.20
pydot==1.4.2
pyenchant==3.2.0
Pygments==2.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.2
pyparsing==2.4.7
pypiwin32==223
pyshorteners==1.0.1
python-dateutil==2.7.5
python-mimeparse==1.6.0
python3-openid==3.2.0
pytz==2020.4
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
sh==1.14.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
spacy==2.3.5
SQLAlchemy==1.2.19
sqlparse==0.2.4
srsly==1.0.5
thinc==7.4.5
toml==0.10.2
tqdm==4.54.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.6
wasabi==0.8.0
watchdog==2.1.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
